Refactor to have both db section in Forest config
