+++ node/db/src/lib.rs
@@ -8,11 +8,13 @@ mod utils;
 
 #[cfg(feature = "rocksdb")]
 pub mod rocks;
-#[cfg(feature = "rocksdb")]
+
 pub mod rocks_config;
 
 #[cfg(feature = "paritydb")]
 pub mod parity_db;
+
+pub mod parity_db_config;
 
 pub use errors::Error;
 pub use memory::MemoryDB;
