+++ utils/statediff/src/main.rs
@@ -46,6 +46,7 @@ fn open_db(chain_path: &Path) -> forest_db::rocks::RocksDb {
 #[cfg(feature = "paritydb")]
 fn open_db(chain_path: &Path) -> forest_db::parity_db::ParityDb {
     use forest_db::parity_db::*;
+    use forest_db::parity_db_config::*;
     ParityDb::open(chain_path.join("paritydb"), &ParityDbConfig::default())
         .expect("Opening ParityDb must succeed")
 }
